Move the Project Header to only the overview page
