+++ frontend/src/pages/project/Overview.vue
@@ -1,5 +1,26 @@
 <template>
-    <div v-if="project.meta" class="space-y-4">
+    <div class="space-y-4">
+        <SectionTopMenu>
+            <template #hero>
+                <div class="h-full flex items-center">
+                    <div class="text-gray-800 text-xl font-bold">
+                        {{ project.name }}
+                    </div>
+                </div>
+            </template>
+            <template #tools>
+                <div class="space-x-2 flex">
+                    <!--  && !isVisitingAdmin -->
+                    <a v-if="editorAvailable" :href="project.url" target="_blank" class="ff-btn ff-btn--secondary" data-action="open-editor">
+                        Open Editor
+                        <span class="ff-btn--icon ff-btn--icon-right">
+                            <ExternalLinkIcon />
+                        </span>
+                    </a>
+                    <DropdownMenu v-if="hasPermission('project:change-status')" buttonClass="ff-btn ff-btn--primary" alt="Open actions menu" :options="options" data-action="open-actions">Actions</DropdownMenu>
+                </div>
+            </template>
+        </SectionTopMenu>
         <div class="grid grid-cols-1 sm:grid-cols-2 gap-4">
             <div class="border rounded p-4">
                 <FormHeading><TemplateIcon class="w-6 h-6 mr-2 inline text-gray-400" />Overview</FormHeading>
@@ -62,43 +83,80 @@
             </div>
         </div>
     </div>
-
 </template>
 
 <script>
-import projectApi from '@/api/project'
-import { ExternalLinkIcon, TrendingUpIcon, TemplateIcon } from '@heroicons/vue/outline'
-import FormHeading from '@/components/FormHeading'
+import { ExternalLinkIcon, TemplateIcon, TrendingUpIcon } from '@heroicons/vue/outline'
+
+import { mapState } from 'vuex'
+
 import ProjectStatusBadge from './components/ProjectStatusBadge'
+
+import projectApi from '@/api/project'
 import AuditLog from '@/components/AuditLog'
+import DropdownMenu from '@/components/DropdownMenu'
+import FormHeading from '@/components/FormHeading'
+import SectionTopMenu from '@/components/SectionTopMenu'
+import permissionsMixin from '@/mixins/Permissions'
 
 export default {
     name: 'ProjectOverview',
-    props: ['project', 'isVisitingAdmin'],
-    computed: {
-        options: function () {
-            return [
-                { name: 'Start', action: async () => { await projectApi.startProject(this.project.id) } },
-                { name: 'Restart', action: async () => { await projectApi.restartProject(this.project.id) } },
-                { name: 'Suspend', action: () => { this.$router.push({ path: `/project/${this.project.id}/settings/danger` }) } }
-            ]
+    components: {
+        AuditLog,
+        DropdownMenu,
+        ExternalLinkIcon,
+        FormHeading,
+        ProjectStatusBadge,
+        SectionTopMenu,
+        TemplateIcon,
+        TrendingUpIcon
+    },
+    mixins: [permissionsMixin],
+    props: {
+        project: {
+            required: true,
+            type: Object
         },
-        editorAvailable: function () {
-            return this.project.meta && this.project.meta.state === 'running'
+        isVisitingAdmin: {
+            required: true,
+            type: Boolean
         }
+    },
+    emits: ['project-start', 'project-delete', 'project-suspend', 'project-restart'],
+    computed: {
+        ...mapState('account', ['teamMembership']),
+        options: function () {
+            const flowActionsDisabled = !(this.project.meta && this.project.meta.state !== 'suspended')
+
+            const result = [
+                {
+                    name: 'Start',
+                    action: () => this.$emit('project-start'),
+                    disabled: this.project.pendingStateChange || this.projectRunning
+                },
+                { name: 'Restart', action: () => this.$emit('project-restart'), disabled: flowActionsDisabled },
+                { name: 'Suspend', class: ['text-red-700'], action: () => this.$emit('project-suspend'), disabled: flowActionsDisabled }
+            ]
+
+            if (this.hasPermission('project:delete')) {
+                result.push(null)
+                result.push({ name: 'Delete', class: ['text-red-700'], action: () => this.$emit('project-delete') })
+            }
+
+            return result
+        },
+        projectRunning () {
+            return this.project?.meta.state === 'running'
+        },
+        editorAvailable () {
+            return this.projectRunning
+        }
+
     },
     methods: {
         loadItems: async function (projectId, cursor) {
             return await projectApi.getProjectAuditLog(projectId, cursor, 4)
         }
-    },
-    components: {
-        ProjectStatusBadge,
-        FormHeading,
-        AuditLog,
-        ExternalLinkIcon,
-        TemplateIcon,
-        TrendingUpIcon
     }
 }
 </script>
